refactor: 기본 컴포넌트 쪼개기 완료(타입 미적용)
