+++ src/components/EditCommentDialog.tsx
@@ -1,0 +1,31 @@
+import { Button } from "../shared/ui/Button"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../shared/ui/Dialog"
+import { Textarea } from "../shared/ui/Textarea"
+
+const EditCommentDialog = ({
+  showEditCommentDialog,
+  setShowEditCommentDialog,
+  selectedComment,
+  setSelectedComment,
+  updateComment,
+}) => {
+  return (
+    <Dialog open={showEditCommentDialog} onOpenChange={setShowEditCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>댓글 수정</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={selectedComment?.body || ""}
+            onChange={(e) => setSelectedComment({ ...selectedComment, body: e.target.value })}
+          />
+          <Button onClick={updateComment}>댓글 업데이트</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
+
+export default EditCommentDialog
